Move eventListeners to View classes
