+++ src/main.js
@@ -1,4 +1,4 @@
-import { render } from './utils/render.js';
+import { render } from './framework/render.js';
 import FilterView from './view/filter-view.js';
 import EventsPresenter from './presenter/events-presenter.js';
 import EventsModel from './model/events-model.js';
