+++ src/view/event-view.js
@@ -71,4 +71,14 @@ export default class EventView extends AbstractView{
   get template() {
     return createEventTemplate(this.#event, this.#offers, this.#destination);
   }
+
+  setEditClickHandler = (cb) => {
+    this._callback.editClick = cb;
+    this.element.querySelector('.event__rollup-btn').addEventListener('click', this.#editClickHandler);
+  };
+
+  #editClickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.editClick();
+  };
 }
